test: co-locate Standard Schema type assertions with runtime tests
