+++ tests/specs/standard-schema/index.ts
@@ -2,5 +2,4 @@ import { describe } from 'manten';
 
 describe('standard-schema', () => {
 	import('./parsing.ts');
-	import('./types.ts');
 });
