+++ tests/specs/standard-schema/parsing.ts
@@ -1,201 +1,303 @@
 import { describe, test, expect } from 'manten';
+import { expectTypeOf } from 'expect-type';
 import * as z from 'zod';
 import * as v from 'valibot';
 import { type } from 'arktype';
 import { typeFlag, getFlag } from '#type-flag';
 
-describe('parsing', () => {
-	describe('Zod', () => {
-		test('enum returns the matched value', () => {
-			const parsed = typeFlag({
+describe('Zod', () => {
+	test('enum returns the matched value', () => {
+		const parsed = typeFlag({
+			size: z.enum(['small', 'medium', 'large']),
+		}, ['--size', 'medium']);
+
+		expect(parsed.flags.size).toBe('medium');
+		expectTypeOf(parsed.flags.size).toEqualTypeOf<'small' | 'medium' | 'large' | undefined>();
+	});
+
+	test('enum rejects an invalid value', () => {
+		let thrown: unknown;
+		try {
+			typeFlag({
 				size: z.enum(['small', 'medium', 'large']),
-			}, ['--size', 'medium']);
+			}, ['--size', 'huge']);
+		} catch (error) {
+			thrown = error;
+		}
 
-			expect(parsed.flags.size).toBe('medium');
-		});
+		// type-flag wraps the thrown message with the flag-name context
+		expect(thrown).toBeInstanceOf(TypeError);
+		expect((thrown as TypeError).message).toMatch('Flag "--size":');
+		expect((thrown as TypeError).message).toMatch('Invalid option');
 
-		test('enum rejects an invalid value', () => {
-			let thrown: unknown;
-			try {
-				typeFlag({
-					size: z.enum(['small', 'medium', 'large']),
-				}, ['--size', 'huge']);
-			} catch (error) {
-				thrown = error;
-			}
+		// the original schema error is preserved on `.cause`
+		expect((thrown as TypeError).cause).toBeInstanceOf(Error);
+		expect(((thrown as TypeError).cause as Error).message).toMatch('Invalid option');
+	});
 
-			// type-flag wraps the thrown message with the flag-name context
-			expect(thrown).toBeInstanceOf(TypeError);
-			expect((thrown as TypeError).message).toMatch('Flag "--size":');
-			expect((thrown as TypeError).message).toMatch('Invalid option');
+	test('coerced number validates the range', () => {
+		const parsed = typeFlag({
+			port: z.coerce.number().int().min(1).max(65_535),
+		}, ['--port', '8080']);
 
-			// the original schema error is preserved on `.cause`
-			expect((thrown as TypeError).cause).toBeInstanceOf(Error);
-			expect(((thrown as TypeError).cause as Error).message).toMatch('Invalid option');
-		});
+		expect(parsed.flags.port).toBe(8080);
+		expectTypeOf(parsed.flags.port).toEqualTypeOf<number | undefined>();
+	});
 
-		test('coerced number validates the range', () => {
-			const parsed = typeFlag({
+	test('coerced number rejects out-of-range values', () => {
+		expect(() => {
+			typeFlag({
 				port: z.coerce.number().int().min(1).max(65_535),
-			}, ['--port', '8080']);
+			}, ['--port', '99999']);
+		}).toThrow('Flag "--port": Too big');
+	});
 
-			expect(parsed.flags.port).toBe(8080);
-		});
+	test('non-coerced number rejects a string value', () => {
+		// The CLI value is always a string, so a plain number schema rejects it
+		expect(() => {
+			typeFlag({
+				port: z.number(),
+			}, ['--port', '3000']);
+		}).toThrow('Flag "--port":');
+	});
 
-		test('coerced number rejects out-of-range values', () => {
-			expect(() => {
-				typeFlag({
-					port: z.coerce.number().int().min(1).max(65_535),
-				}, ['--port', '99999']);
-			}).toThrow('Flag "--port": Too big');
-		});
+	test('transform maps the value', () => {
+		const parsed = typeFlag({
+			list: z.string().transform(value => value.split(',')),
+		}, ['--list', 'a,b,c']);
 
-		test('non-coerced number rejects a string value', () => {
-			// The CLI value is always a string, so a plain number schema rejects it
-			expect(() => {
-				typeFlag({
-					port: z.number(),
-				}, ['--port', '3000']);
-			}).toThrow('Flag "--port":');
-		});
+		expect(parsed.flags.list).toStrictEqual(['a', 'b', 'c']);
+		expectTypeOf(parsed.flags.list).toEqualTypeOf<string[] | undefined>();
+	});
 
-		test('transform maps the value', () => {
-			const parsed = typeFlag({
-				list: z.string().transform(value => value.split(',')),
-			}, ['--list', 'a,b,c']);
+	test('repeated flag collects and validates each element', () => {
+		const parsed = typeFlag({
+			tag: [z.enum(['debug', 'info', 'warn'])],
+		}, ['--tag', 'debug', '--tag', 'warn']);
 
-			expect(parsed.flags.list).toStrictEqual(['a', 'b', 'c']);
-		});
+		expect(parsed.flags.tag).toStrictEqual(['debug', 'warn']);
+		expectTypeOf(parsed.flags.tag).toEqualTypeOf<('debug' | 'info' | 'warn')[]>();
+	});
 
-		test('repeated flag collects and validates each element', () => {
-			const parsed = typeFlag({
+	test('repeated flag rejects an invalid element', () => {
+		expect(() => {
+			typeFlag({
 				tag: [z.enum(['debug', 'info', 'warn'])],
-			}, ['--tag', 'debug', '--tag', 'warn']);
-
-			expect(parsed.flags.tag).toStrictEqual(['debug', 'warn']);
-		});
-
-		test('repeated flag rejects an invalid element', () => {
-			expect(() => {
-				typeFlag({
-					tag: [z.enum(['debug', 'info', 'warn'])],
-				}, ['--tag', 'debug', '--tag', 'nope']);
-			}).toThrow('Flag "--tag": Invalid option');
-		});
-
-		test('type-flag default applies when the flag is absent', () => {
-			const parsed = typeFlag({
-				size: {
-					type: z.enum(['small', 'large']),
-					default: 'small',
-				},
-			}, []);
-
-			expect(parsed.flags.size).toBe('small');
-		});
-
-		test('absent raw schema flag is undefined (schema .default is not a type-flag default)', () => {
-			const parsed = typeFlag({
-				size: z.enum(['small', 'large']),
-			}, []);
-
-			expect(parsed.flags.size).toBe(undefined);
-		});
+			}, ['--tag', 'debug', '--tag', 'nope']);
+		}).toThrow('Flag "--tag": Invalid option');
 	});
 
-	// Proves the adapter is library-agnostic, not coupled to Zod
-	describe('Valibot', () => {
-		test('picklist returns the matched value', () => {
-			const parsed = typeFlag({
+	test('type-flag default applies when the flag is absent', () => {
+		const parsed = typeFlag({
+			size: {
+				type: z.enum(['small', 'large']),
+
+				// `as const` preserves the literal union; a plain default widens to string
+				default: 'small' as const,
+			},
+		}, []);
+
+		expect(parsed.flags.size).toBe('small');
+		expectTypeOf(parsed.flags.size).toEqualTypeOf<'small' | 'large'>();
+	});
+
+	test('absent raw schema flag is undefined (schema .default is not a type-flag default)', () => {
+		const parsed = typeFlag({
+			size: z.enum(['small', 'large']),
+		}, []);
+
+		expect(parsed.flags.size).toBe(undefined);
+		expectTypeOf(parsed.flags.size).toEqualTypeOf<'small' | 'large' | undefined>();
+	});
+
+	test('{ type } object form behaves like the bare schema', () => {
+		const parsed = typeFlag({
+			size: {
+				type: z.enum(['small', 'large']),
+			},
+		}, ['--size', 'small']);
+
+		expect(parsed.flags.size).toBe('small');
+		expectTypeOf(parsed.flags.size).toEqualTypeOf<'small' | 'large' | undefined>();
+	});
+
+	test('{ type: [schema] } object form collects an array', () => {
+		const parsed = typeFlag({
+			tags: {
+				type: [z.string()],
+			},
+		}, ['--tags', 'a', '--tags', 'b']);
+
+		expect(parsed.flags.tags).toStrictEqual(['a', 'b']);
+		expectTypeOf(parsed.flags.tags).toEqualTypeOf<string[]>();
+	});
+
+	test('array schema with a default falls back to the default and stays a clean array', () => {
+		const parsed = typeFlag({
+			tags: {
+				type: [z.string()],
+				default: () => ['x'],
+			},
+		}, []);
+
+		expect(parsed.flags.tags).toStrictEqual(['x']);
+
+		// Guards against the default unioning into an un-reduced `string[] | string[]`
+		expectTypeOf(parsed.flags.tags).toEqualTypeOf<string[]>();
+	});
+});
+
+// Proves the adapter is library-agnostic, not coupled to Zod
+describe('Valibot', () => {
+	test('picklist returns the matched value', () => {
+		const parsed = typeFlag({
+			mode: v.picklist(['dev', 'prod']),
+		}, ['--mode', 'dev']);
+
+		expect(parsed.flags.mode).toBe('dev');
+		expectTypeOf(parsed.flags.mode).toEqualTypeOf<'dev' | 'prod' | undefined>();
+	});
+
+	test('picklist rejects an invalid value', () => {
+		expect(() => {
+			typeFlag({
 				mode: v.picklist(['dev', 'prod']),
-			}, ['--mode', 'dev']);
+			}, ['--mode', 'staging']);
+		}).toThrow('Flag "--mode": Invalid type');
+	});
 
-			expect(parsed.flags.mode).toBe('dev');
-		});
+	test('email validates the string', () => {
+		const parsed = typeFlag({
+			email: v.pipe(v.string(), v.email()),
+		}, ['--email', 'user@example.com']);
 
-		test('picklist rejects an invalid value', () => {
-			expect(() => {
-				typeFlag({
-					mode: v.picklist(['dev', 'prod']),
-				}, ['--mode', 'staging']);
-			}).toThrow('Flag "--mode": Invalid type');
-		});
+		expect(parsed.flags.email).toBe('user@example.com');
+		expectTypeOf(parsed.flags.email).toEqualTypeOf<string | undefined>();
+	});
 
-		test('email validates the string', () => {
-			const parsed = typeFlag({
+	test('email rejects an invalid string', () => {
+		expect(() => {
+			typeFlag({
 				email: v.pipe(v.string(), v.email()),
-			}, ['--email', 'user@example.com']);
+			}, ['--email', 'not-an-email']);
+		}).toThrow('Flag "--email": Invalid email');
+	});
+});
 
-			expect(parsed.flags.email).toBe('user@example.com');
-		});
+// ArkType schemas are callable (typeof === 'function') yet implement Standard
+// Schema, so they must still route through validation, not the raw-parser path.
+describe('ArkType (callable schema)', () => {
+	test('returns the matched value', () => {
+		const parsed = typeFlag({
+			size: type("'small' | 'large'"),
+		}, ['--size', 'small']);
 
-		test('email rejects an invalid string', () => {
-			expect(() => {
-				typeFlag({
-					email: v.pipe(v.string(), v.email()),
-				}, ['--email', 'not-an-email']);
-			}).toThrow('Flag "--email": Invalid email');
-		});
+		expect(parsed.flags.size).toBe('small');
+		expectTypeOf(parsed.flags.size).toEqualTypeOf<'small' | 'large' | undefined>();
 	});
 
-	// ArkType schemas are callable (typeof === 'function') yet implement Standard
-	// Schema, so they must still route through validation, not the raw-parser path.
-	describe('ArkType (callable schema)', () => {
-		test('returns the matched value', () => {
-			const parsed = typeFlag({
+	test('rejects an invalid value', () => {
+		expect(() => {
+			typeFlag({
 				size: type("'small' | 'large'"),
-			}, ['--size', 'small']);
-
-			expect(parsed.flags.size).toBe('small');
-		});
-
-		test('rejects an invalid value', () => {
-			expect(() => {
-				typeFlag({
-					size: type("'small' | 'large'"),
-				}, ['--size', 'huge']);
-			}).toThrow('Flag "--size":');
-		});
-
-		test('absent flag is undefined (callable schema is not treated as a default)', () => {
-			const parsed = typeFlag({
-				size: type("'small' | 'large'"),
-			}, []);
-
-			expect(parsed.flags.size).toBe(undefined);
-		});
+			}, ['--size', 'huge']);
+		}).toThrow('Flag "--size":');
 	});
 
-	describe('getFlag', () => {
-		test('parses and validates with a schema', () => {
-			const size = getFlag('--size', z.enum(['small', 'large']), ['--size', 'small']);
+	test('absent flag is undefined (callable schema is not treated as a default)', () => {
+		const parsed = typeFlag({
+			size: type("'small' | 'large'"),
+		}, []);
 
-			expect(size).toBe('small');
-		});
+		expect(parsed.flags.size).toBe(undefined);
+	});
+});
 
-		test('rejects an invalid value', () => {
-			expect(() => {
-				getFlag('--size', z.enum(['small', 'large']), ['--size', 'huge']);
-			}).toThrow('Flag "--size": Invalid option');
-		});
+describe('getFlag', () => {
+	test('parses and validates with a schema', () => {
+		const size = getFlag('--size', z.enum(['small', 'large']), ['--size', 'small']);
+
+		expect(size).toBe('small');
+		expectTypeOf(size).toEqualTypeOf<'small' | 'large' | undefined>();
 	});
 
-	describe('Async', () => {
-		test('async validation throws (parsing is synchronous)', () => {
-			// Hand-rolled Standard Schema with an async `validate`; no dependency needed
-			const asyncSchema = {
-				'~standard': {
-					version: 1 as const,
-					vendor: 'test',
-					validate: () => Promise.resolve({ value: 'async' }),
-				},
-			};
+	test('rejects an invalid value', () => {
+		expect(() => {
+			getFlag('--size', z.enum(['small', 'large']), ['--size', 'huge']);
+		}).toThrow('Flag "--size": Invalid option');
+	});
 
-			expect(() => {
-				typeFlag({
-					asyncFlag: asyncSchema,
-				}, ['--async-flag', 'value']);
-			}).toThrow('Flag "--async-flag": Async schema validation is not supported');
-		});
+	test('collects an array from a wrapped schema', () => {
+		const tags = getFlag('--tag', [z.string()], ['--tag', 'a', '--tag', 'b']);
+
+		expect(tags).toStrictEqual(['a', 'b']);
+		expectTypeOf(tags).toEqualTypeOf<string[]>();
+	});
+});
+
+describe('Async', () => {
+	test('async validation throws (parsing is synchronous)', () => {
+		// Hand-rolled Standard Schema with an async `validate`; no dependency needed
+		const asyncSchema = {
+			'~standard': {
+				version: 1 as const,
+				vendor: 'test',
+				validate: () => Promise.resolve({ value: 'async' }),
+			},
+		};
+
+		expect(() => {
+			typeFlag({
+				asyncFlag: asyncSchema,
+			}, ['--async-flag', 'value']);
+		}).toThrow('Flag "--async-flag": Async schema validation is not supported');
+	});
+});
+
+describe('Inference', () => {
+	test('schema and native flags infer independently in one call', () => {
+		const parsed = typeFlag({
+			schemaEnum: z.enum(['a', 'b']),
+			nativeBoolean: Boolean,
+			nativeString: String,
+			schemaArray: [z.coerce.number()],
+			schemaDefault: {
+				type: z.enum(['x', 'y']),
+				default: 'x' as const,
+			},
+		}, ['--schema-enum', 'a', '--native-boolean', '--native-string', 'hi', '--schema-array', '3']);
+
+		expect(parsed.flags.schemaEnum).toBe('a');
+		expect(parsed.flags.nativeBoolean).toBe(true);
+		expect(parsed.flags.schemaArray).toStrictEqual([3]);
+		expect(parsed.flags.schemaDefault).toBe('x');
+
+		expectTypeOf(parsed.flags).toEqualTypeOf<{
+			schemaEnum: 'a' | 'b' | undefined;
+			nativeBoolean: boolean | undefined;
+			nativeString: string | undefined;
+			schemaArray: number[];
+			schemaDefault: 'x' | 'y';
+		}>();
+	});
+
+	test('a schema whose output is an array is treated as a scalar', () => {
+		// Documented gotcha: `z.array(...)` validates a single token, so use `[schema]`
+		// to accept multiple. At runtime it rejects the token; the type is the output.
+		const schemas = {
+			tags: z.array(z.string()),
+		};
+
+		expect(() => typeFlag(schemas, ['--tags', 'a'])).toThrow('Flag "--tags":');
+
+		const parsed = typeFlag(schemas, []);
+		expectTypeOf(parsed.flags.tags).toEqualTypeOf<string[] | undefined>();
+	});
+
+	test('rejects a multi-element schema tuple', () => {
+		typeFlag({
+			// @ts-expect-error only one element is allowed in the array form
+			tags: [z.string(), z.string()],
+		}, []);
 	});
 });
